Created a package for each screen
